Clean up code; extract to components
