+++ apps/web/src/components/AutoCompleteInput/AutoCompleteInput.tsx
@@ -30,7 +30,7 @@ export const AutoCompleteInput = ({
       !inputRef.current ||
       !window.google.maps?.places
     ) {
-      console.warn("google");
+      console.warn("google maps not loaded yet");
       return;
     }
 
@@ -53,17 +53,22 @@ export const AutoCompleteInput = ({
       }
 
       geocoder.geocode({ location }, (results, status) => {
-        if (status === google.maps.GeocoderStatus.OK && results && results[0]) {
-          const addressComponents = results[0].address_components;
-          console.log(addressComponents);
-
-          const adr = extractAddressComponents(addressComponents);
-          console.log(adr);
-          onPlaceResolved?.(place, addressComponents);
-        } else {
+        if (
+          status !== google.maps.GeocoderStatus.OK ||
+          !results ||
+          !results[0]
+        ) {
           toast.error("Error getting location");
           console.error("Geocoder failed:", status);
+          return;
         }
+
+        const addressComponents = results[0].address_components;
+        console.log(addressComponents);
+
+        const adr = extractAddressComponents(addressComponents);
+        console.log(adr);
+        onPlaceResolved?.(place, addressComponents);
       });
     });
